# docs/swagger.yaml
basePath: /
definitions:
...
  adminreq.UserCreateRequest:
    properties:
      nickname:
        maxLength: 64
        type: string
      password:
        maxLength: 64
        minLength: 6
        type: string
      role:
        enum:
        - admin
        - user
        type: string
      username:
        maxLength: 32
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
  adminreq.UserUpdateRequest:
    properties:
      nickname:
        maxLength: 64
        type: string
      password:
        maxLength: 64
        minLength: 6
        type: string
      role:
        enum:
        - admin
        - user
        type: string
    type: object
...
  response.Body:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
      request_id:
        type: string
      trace_id:
        type: string
    type: object
info:
  contact: {}
  description: 企业级 Gin 脚手架示例 API
  title: Gin Scaffold API
  version: "1.0"
paths:
...
  /api/v1/admin/users:
    get:
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页条数
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
      summary: 用户列表（后台）
      tags:
      - admin-user
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/adminreq.UserCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
      summary: 创建用户（后台）
      tags:
      - admin-user
  /api/v1/admin/users/{id}:
    delete:
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
      summary: 删除用户（后台）
      tags:
      - admin-user
    get:
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
      summary: 用户详情（后台）
      tags:
      - admin-user
    put:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/adminreq.UserUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
      summary: 更新用户（后台）
      tags:
      - admin-user
  /api/v1/admin/users/export/tasks:
    post:
      parameters:
      - description: 用户名模糊查询
        in: query
        name: username
        type: string
      - description: 昵称模糊查询
        in: query
        name: nickname
        type: string
      - description: '导出列，逗号分隔: id,username,nickname,created_at,role'
        in: query
        name: fields
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
      summary: 用户异步导出任务创建（后台）
      tags:
      - admin-user
  /api/v1/admin/users/export/tasks/{task_id}:
    get:
      parameters:
      - description: 任务ID
        in: path
        name: task_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
      summary: 用户异步导出任务状态（后台）
      tags:
      - admin-user
  /api/v1/admin/users/export/tasks/{task_id}/download:
    get:
      parameters:
      - description: 任务ID
        in: path
        name: task_id
        required: true
        type: string
      produces:
      - text/csv
      responses:
        "200":
          description: OK
          schema:
            type: file
      summary: 下载用户异步导出结果（后台）
      tags:
      - admin-user
...
swagger: "2.0"